requirements.txt
sentence-transformers==2.7.0
chromadb==0.5.5
openai==1.40.6
httpx==0.27.0
python-dotenv==1.0.1